Finalizando API Doador

+++ Doador.Domain/Commands/DoadorCommand.cs
@@ -6,7 +6,7 @@
         public string DoadorNome { get; set; }
         public string DoadorCidade { get; set; }
         public string DoadorEstado { get; set; }
-        public long DoadorCEP { get; set; }
+        public int DoadorCEP { get; set; }
         public string DoadorEmail { get; set; }
         public int DoadorTelefone { get; set; }
     }
